Show more descriptive dependency paths
When issue dependency paths are shown, they may be sourced from
multiple projects (either in the CLI or in GitHub, depending on
how the tool is used). Multiple projects may have different
prefixes to denote that they are sources from a different code
branch. This commit intelligently shows that prefix in the
dependency paths when necessary.