Remove unused UUID validation code
This was left over from a previous refactoring.

The added complexity of validating the UUIDs with this library was too
large due to how the `enquirer` module works, so it was never properly
implemented.

There's really no added benefit of validating the UUIDs as the validator
only validates the UUID format, not that the UUIDs exists on the remote
server. If they do not exist, we'll know soon enough anyway.